Moved FileSystemHandler to namespace Utils
