+++ gameoflife.php
@@ -11,7 +11,7 @@ $loader->addPsr4("GameOfLife\\", __DIR__ . "/src/GameOfLife");
 $loader->addPsr4("Ulrichsg\\", __DIR__ . "/src/Ulrichsg");
 $loader->addPsr4("Input\\", __DIR__ . "/src/GameOfLife/Inputs");
 $loader->addPsr4("Output\\", __DIR__ . "/src/GameOfLife/Outputs");
-
+$loader->addPsr4("Utils\\", __DIR__ . "/src/GameOfLife/Utils");
 
 use GameOfLife\Board;
 use GameOfLife\RuleSet;
