+++ src/GameOfLife/Utils/FileSystemHandler.php
@@ -6,7 +6,7 @@
  * @author Yannick Lapp <yannick.lapp@cn-consult.eu>
  */
 
-namespace GameOfLife;
+namespace Utils;
 
 /**
  * Class FileSystemHandler
