+++ tests/Outputs/VideoOutputTest.php
@@ -10,7 +10,7 @@ use Output\VideoOutput;
 use PHPUnit\Framework\TestCase;
 use GameOfLife\Board;
 use GameOfLife\RuleSet;
-use GameOfLife\FileSystemHandler;
+use Utils\FileSystemHandler;
 use Ulrichsg\Getopt;
 
 /**
